+++ minijinja/src/vm/loop_object.rs
@@ -10,6 +10,8 @@ pub(crate) struct Loop {
     pub len: usize,
     pub idx: AtomicUsize,
     pub depth: usize,
+    #[cfg(feature = "adjacent-loop-items")]
+    pub value_triple: Mutex<(Option<Value>, Option<Value>, Option<Value>)>,
     pub last_changed_value: Mutex<Option<Vec<Value>>>,
 }
 
@@ -74,6 +76,10 @@ impl StructObject for Loop {
                 "last",
                 "depth",
                 "depth0",
+                #[cfg(feature = "adjacent-loop-items")]
+                "previtem",
+                #[cfg(feature = "adjacent-loop-items")]
+                "nextitem",
             ][..],
         )
     }
@@ -97,6 +103,24 @@ impl StructObject for Loop {
             "last" => Some(Value::from(len == 0 || idx == len - 1)),
             "depth" => Some(Value::from(self.depth + 1)),
             "depth0" => Some(Value::from(self.depth)),
+            #[cfg(feature = "adjacent-loop-items")]
+            "previtem" => Some(
+                self.value_triple
+                    .lock()
+                    .unwrap()
+                    .0
+                    .clone()
+                    .unwrap_or(Value::UNDEFINED),
+            ),
+            #[cfg(feature = "adjacent-loop-items")]
+            "nextitem" => Some(
+                self.value_triple
+                    .lock()
+                    .unwrap()
+                    .2
+                    .clone()
+                    .unwrap_or(Value::UNDEFINED),
+            ),
             _ => None,
         }
     }